requirements.txt
python-telegram-bot==13.14
numpy==1.23.4
opencv-python==4.6.0.66
torch==1.12.1
pandas==1.5.0
requests==2.28.1
Pillow==9.2.0
torchvision==0.13.1
PyYAML==6.0
tqdm==4.64.1
matplotlib==3.6.1
seaborn==0.12.0